[FEAT] : QuoteView 예외 사항 메서드 생성 (remove, replace, semaphore) (#34)

+++ FakeBithumbAssignment/FakeBithumbAssignment/Network/OrderbookAPIResponse.swift
@@ -25,7 +25,7 @@ struct OrderbookAPIResponse: Codable {
 struct Quote: Codable {
     let price, quantity: String
     
-    init(price: Double, quantity: Double) {
+    init(price: Int, quantity: Double) {
         self.price = "\(price)"
         self.quantity = "\(quantity)"
     }
